Lammps implement file interface
